+++ R/hardCode.R
@@ -1,7 +1,8 @@
 ##' Facilitate hard coding constants into R
 ##'
 ##' Hard coding isn't the best practice, but sometimes it's useful, especially in
-##' one-off scripts for analyses.  This
+##' one-off scripts for analyses. An typical example would be to select a large
+##' number of columns in a dataset by their names.  This
 ##' function facilitate hard coding constants into R by printing the code from a
 ##' vector that would be needed to create that vector.
 ##'