+++ R/interactionPlot.R
@@ -21,10 +21,11 @@
 ##'
 ##' Plots the mean (or other summary) of the response for two-way combinations
 ##' of factors, thereby illustrating possible interactions.
-## %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+## %%% BEGIN %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 ##' This modifies \code{\link{interaction.plot}} by adding the \code{errorBar} and
-##' \code{jitterErrorBars} arguments.
-## %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+##' \code{jitterErrorBars} arguments, making it possible to add error bars of the
+##' same length to the plot.
+## %%% END %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 ##'
 ##' @details By default the levels of \code{x.factor} are plotted on the x axis in their
 ##' given order, with extra space left at the right for the legend (if
@@ -42,34 +43,30 @@
 ##' 1:9, 0, and the capital letters to plot the traces.
 ##'
 ##' @export
-## @usage interactionPlot(x.factor, trace.factor, response, fun = mean,
-##                        errorBar = NULL, jitterErrorBars = FALSE,
-##	                       type = c("l", "p", "b", "o", "c"), legend = TRUE,
-##                        trace.label = deparse(substitute(trace.factor)), fixed = FALSE,
-##                        xlab = deparse(substitute(x.factor)), ylab = ylabel,
-##                        ylim = range(cells, na.rm=TRUE),
-##                        lty = nc:1, col = 1, pch = c(1L:9, 0, letters),
-##                        xpd = NULL, leg.bg = par("bg"), leg.bty = "n",
-##                        xtick = FALSE, xaxt = par("xaxt"), axes = TRUE, ...)
-##
 ##' @param x.factor a factor whose levels will form the x axis.
 ##' 
 ##' @param trace.factor another factor whose levels will form the traces.
 ##' 
-##' @param response a numeric variable giving the response
+##' @param response a numeric variable giving the response.
 ##' 
 ##' @param fun the function to compute the summary. Should return a single real
 ##' value.
 ##' 
-##' @param errorBar A list with required elements \code{height} and
+## %%% BEGIN %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+##'
+##' @param errorBar A list with required elements \code{barLength} and
 ##' \code{width}, and optional element \code{blankMiddle}. These correspond to
 ##' the arguments of \code{\link{vertErrorBar}}.  Defaults to \code{NULL}, in
-##' which case no error bars are drawn
+##' which case no error bars are drawn.  Currently, error bars must all be the same
+##' length and have the same width. Thus, elements in the list should be atomic (of length 1).
 ##' 
-##' @param jitterErrorBars Logical indicating whether the x-values of the points
-##' (and the error bars) should be jittered for easier readability. This
-##' argument is ignored if \code{errorBar = NULL}.
+##' @param jitterErrorBars A list with the \code{factor} and/or \code{amount} arguments that will be
+##' passed to \code{\link{jitter}}, indicating the amount (and/or factor) by which the x-values of the points
+##' (and the error bars) should be jittered for easier readibility. No jittering is performed if
+##' \code{jitterErrorBars = NULL}. This argument is ignored if \code{errorBar = NULL}.
 ##' 
+## %%% END %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+##'
 ##' @param type the type of plot (see \code{\link{plot.default}}): lines or
 ##' points or both.
 ##' 
@@ -133,9 +130,9 @@
 ##'
 ##' # Add error bars
 ##' interactionPlot(dose, supp, len, fixed = TRUE, col  =  2:3, type = "p",
-##'                 errorBar = list(height = 3, width = 0.05, blankMiddle = 1),
-##'                 jitterErrorBars = TRUE,
-##'                 main = "Error bars have no meaning--just for illustration")
+##'                 errorBar = list(barLength = 3, width = 0.05, blankMiddle = 1),
+##'                 jitterErrorBars = list(factor = 1.5),
+##'                 main = "Error bars have no meaning: just for illustration")
 ##' })
 ## %%% END %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -145,7 +142,7 @@ interactionPlot <-
 ## BEGIN ##############################################################################
 
              errorBar = NULL,
-             jitterErrorBars = FALSE,
+             jitterErrorBars = NULL,
 
 ## END ################################################################################
 
@@ -195,27 +192,47 @@ interactionPlot <-
     if (!is.null(errorBar)) {
 
       # Check list
-      stopifnot(is.list(errorBar))
-
-      if (!all(names(errorBar) %in% c("height", "width", "blankMiddle")))
-        stop("'errorBar' list must only have names 'height', 'width', and optionally, 'blankMiddle'")
-
-      if (!all(c("height", "width") %in% names(errorBar)))
-        stop("'errorBar' list must have names 'height', 'width', and optionally, 'blankMiddle'")
+      stopifnotMsg(# Must be a list
+                   if (is.list(errorBar)) {
+                     # May only have 'barLength', 'width', and/or 'blankMiddle'
+                     all(names(errorBar) %in% c("barLength", "width", "blankMiddle"))  &
+                     # Must have 'barLength' and 'width'
+                     all(c("barLength", "width") %in% names(errorBar)) &
+                     # All args must be numeric
+                     all(unlist(lapply(errorBar, is.numeric))) &
+                     # Args must be of length 1
+                     all(lengths(errorBar) == 1)
+                   } else FALSE,
+                   "'errorBar' must be a list of atomic, numeric elements with names 'barLength', 'width', and optionally, 'blankMiddle'")
 
       # Change the ylim so error bars will not be chopped off if the user didn't change the defaults
-      if (!ylimChanged)
-        ylim <- ylim + c(- errorBar$height / 2, errorBar$height / 2)
+      if (!ylimChanged) {
+        ylim <- ylim + c(- errorBar$barLength / 2, errorBar$barLength / 2)
+      }
 
       # Rearrange xvals a bit if jitter is requested
-      if (jitterErrorBars) {
+      if (!is.null(jitterErrorBars)) {
+
+        stopifnotMsg(# Argument must be a list
+                     if (is.list(jitterErrorBars)) {
+                       # No more than length 2
+                       (length(jitterErrorBars) <= 2) &
+                       # Names must be "amount" and/or "factor"
+                       all(names(jitterErrorBars) %in% c("amount", "factor")) &
+                       # Both args must be numeric
+                       all(unlist(lapply(jitterErrorBars, is.numeric))) &
+                       # All must be of length 1
+                       all(lengths(jitterErrorBars) == 1)
+                     } else FALSE,
+                     "'jitterErrorBars' must be a list of atomic, numeric arguments with names 'amount' and/or 'factor'")
 
         xvals_orig <- xvals
 
         xvals <- matrix(NA, nrow = NROW(cells), ncol = NCOL(cells))
 
-        for (i in 1:NCOL(xvals))
-          xvals[,i] <- jitter(xvals_orig)
+        for (i in 1:NCOL(xvals)) {
+          xvals[,i] <- do.call("jitter", c(list(x = xvals_orig), jitterErrorBars))
+        }
 
       }
 
@@ -232,21 +249,25 @@ interactionPlot <-
     # Add in error bars
     if (!is.null(errorBar)) {
 
-      if (!jitterErrorBars) {
+      # If no jittering
+      if (is.null(jitterErrorBars)) {
 
         for (cn in 1:NCOL(cells)) {
-          for (xv in 1:length(xvals))
-            do.call("vertErrorBar", c(list(x = xvals[xv], center = cells[xv,cn]), errorBar))
+          for (xv in 1:length(xvals)) {
+            do.call("vertErrorBar", c(list(x = xvals[xv], center = cells[xv, cn]), errorBar))
+          }
         }
 
       }
+      # If jittering
       else {
 
         cn <- as.vector(cells)
         xv <- as.vector(xvals)
 
-        for (i in 1:length(cn))
+        for (i in 1:length(cn)) {
           do.call("vertErrorBar", c(list(x = xv[i], center = cn[i]), errorBar))
+        }
 
         # Restore xvals to it's original state
         xvals <- xvals_orig